extract useListenSmp, layout
